Part1V5.py
Put creation of single burglary zone into a while loop so that it can iterate for set number of generations. This mimics the ‘jump’ of a burglar to another location as once the while loop has been run through, it will begin again at a different location. Added in print out to show progress through iterations. Made program delete shapefiles for storing initial and subsequent points and initial buffer. This was so that the file names could be reused when looped.