Part3V2.py
Created global point variable to store the last initial burglary point. Adapted code from http://gis.stackexchange.com/questions/76077/creating-points-based-on-distance-and-bearing-from-survey-point-using-qgis to take a distance figure and bearing angle (currently set to 0) to calculate the X,Y values of the next burglary point. Integrated this into a new function called specificPoint to calculate the next jump in the sequence. Created new feature class for this point and added it to burglary.shp as before, storing this new point as the next prevPoint value.