Part3V3.py
Used Pythons random number generator to generate a number between 0 and 360 for each bearing so that the burglar moves around at random. Got the boundary feature using a search cursor and used the ‘.contains’ operator to check if the new point lay in the boundary. Ran code to use the point as normal if true, otherwise disregard point and call function again with a new bearing parameter.